refactor: add routes/tasks.py
